Move scheduling queue's nominator to a separate file
